+++ src/dialects/mssql/transaction.js
@@ -33,6 +33,10 @@ module.exports = class Transaction_MSSQL extends Transaction {
       () => {
         let err = error;
         if (isUndefined(error)) {
+          if (this.doNotRejectOnRollback) {
+            this._resolver();
+            return;
+          }
           err = new Error(`Transaction rejected with non-error: ${error}`);
         }
         this._rejecter(err);