+++ src/dialects/mysql/transaction.js
@@ -30,6 +30,14 @@ assign(Transaction_MySQL.prototype, {
         if (status === 1) t._resolver(value);
         if (status === 2) {
           if (isUndefined(value)) {
+            if (
+              sql &&
+              sql.toUpperCase() === 'ROLLBACK' &&
+              t.doNotRejectOnRollback
+            ) {
+              t._resolver();
+              return;
+            }
             value = new Error(`Transaction rejected with non-error: ${value}`);
           }
           t._rejecter(value);